requirements.txt
python-telegram-bot
requests
sqlalchemy==2.0.18
python-dotenv>=1.0.0
tabulate